kernel - loadavg structure 32->64 bit fields
* The loadavg structure overflows when a large number of processes
  are running.  Yes, I in fact got it to overflow.  Change the load
  fields from 32 to 64 bits.

* Tested to 400,000 runnable processes.

+++ sys/kern/kern_synch.c
@@ -1235,7 +1235,7 @@ loadav(void *arg)
 	avg = &averunnable;
 	for (i = 0; i < 3; i++) {
 		avg->ldavg[i] = (cexp[i] * avg->ldavg[i] +
-		    nrun * FSCALE * (FSCALE - cexp[i])) >> FSHIFT;
+		    (long)nrun * FSCALE * (FSCALE - cexp[i])) >> FSHIFT;
 	}
 
 	/*
